feat!: actually initial commit
Implemented a major overhaul on top of old code started in 2024:

1. Alloc and dependency free protocol crate.
2. Cli whose main purpose is to test protocol crate but will probably
   develope over time as a useful tool for debugging TPM chips.

Based on my earlier work https://github.com/ColinIanKing/tpm2-scripts in
terms of functionality, and is continuation for that work. `tpm2-scripts`
was developed to test run TPM2 when the kernel gained that support and
later forked and refined as the kselftest.

Similarly `tpm2_protocol` targets as being basis for `linux-rust` some day
in foreseeable future, as using using type system for high fidely protocol
processing is the measurable "extra gain" that can be acquired from using
Rust.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>