fix(protocol): tpm_parse_command() handle processing
`tpm_parse_command()` failed  with zero handles because the buffer was
not advanced correctly past the header, causing parameter data to be
misinterpreted as a session block.

A related bug was fixed in the `tpm_parse_response()`, which did not
correctly interpret parameter size in

The bugs were root caused by refining the tests to provide more detail,
and adding more command parsing tests.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>