refactor!(tpm2_protocol): move dyn trait out of crate
TpmObject created a confusing API because it was not used by the library's
primary parsing functions, which rely on enums. Thus, dynamic dispatch
pattern that was disconnected from the library's core design.

Remove the TpmObject trait from the tpm2_protocol library.

The dyn trait shenanigns are migrated to dynamic round-trip test, which
demonstrate how to do the "dynamic decoration" properly, and perform "blind
parsing" despite the core crate does not have the "dynamic ability"

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>