fix(tpm2sh): pipeline fixes
The `pcr-read` command was not participating in the command pipeline
correctly. It did not read from stdin, causing it to drop any objects that
were piped to it.

The `print-stack` command was printing pipeline objects in a FIFO
(first-in, first-out) order and was adding a spurious separator line
between each object.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>